add checks client is operationnal

+++ src/Manager/DocumentManager.php
@@ -7,7 +7,7 @@ use ACSEO\TypesenseBundle\Client\TypesenseClient;
 class DocumentManager
 {
     private $client;
-    
+
     public function __construct(TypesenseClient $client)
     {
         $this->client = $client;
@@ -15,16 +15,28 @@ class DocumentManager
 
     public function delete($collection, $id)
     {
+        if (!$this->client->isOperationnal()) {
+            return null;
+        }
+
         return $this->client->collections[$collection]->documents[$id]->delete();
     }
-    
+
     public function index($collection, $data)
     {
+        if (!$this->client->isOperationnal()) {
+            return null;
+        }
+
         return $this->client->collections[$collection]->documents->create($data);
     }
 
     public function import(string $collection, array $data, string $action = 'create')
     {
+        if (!$this->client->isOperationnal()) {
+            return null;
+        }
+
         return $this->client->collections[$collection]->documents->import($data, ['action' => $action]);
     }
 }
